use new heartbeater across project
