+++ src/stackdriver-nozzle/cmd/firehose_rate.go
@@ -28,6 +28,9 @@ import (
 )
 
 func main() {
+	logger := lager.NewLogger("firehose-rate-script")
+	logger.RegisterSink(lager.NewWriterSink(os.Stdout, lager.DEBUG))
+
 	apiEndpoint := os.Getenv("FIREHOSE_ENDPOINT")
 	username := os.Getenv("FIREHOSE_USERNAME")
 	password := os.Getenv("FIREHOSE_PASSWORD")
@@ -39,16 +42,16 @@ func main() {
 		Password:          password,
 		SkipSslValidation: skipSSLValidation}
 
-	cfClient := cfclient.NewClient(cfConfig)
-
+	cfClient, err := cfclient.NewClient(cfConfig)
+	if err != nil {
+		logger.Fatal("NewClient", err)
+	}
 	client := cloudfoundry.NewFirehose(cfConfig, cfClient, "firehose-rate-script")
 
 	messages, _ := client.Connect()
 
-	trigger := time.Tick(1 * time.Second)
-	logger := lager.NewLogger("firehose-rate-script")
-	logger.RegisterSink(lager.NewWriterSink(os.Stdout, lager.DEBUG))
-	heartbeater := heartbeat.NewHeartbeater(logger, trigger)
+	period := time.Duration(1 * time.Second)
+	heartbeater := heartbeat.NewTelemetry(logger, period, nil)
 	heartbeater.Start()
 	defer heartbeater.Stop()
 	for _ = range messages {
